Create BlackAndWhiteSnakeGame.py
This game uses a simple concept where the snake is defined using a head, a tail and a body. The snake is regulated using the arrow keys. The screen contains an invisible boundary all around with which if the snake collides, the game ends. also the snake upon colliding with itself ends the game with an error and a message, stating the reason for ending the game. The snake enlarges upon hitting each target food.

NOTE: The game runs only with those versions that supports curses. Alternative for "curses" is "urwid".